Playing around with CSS.
Going through and trying to learn the basics of flexbox, grid, bootstrap and media queries. So far I've got grid down (and media queries are quite easy and intuitive). So going to set a goal of finishing up bootstrap and flexbox by end of day(aka, by building a quick project with them).